Adding license information.
The README was updated with license information. We released the code
under the GPL, and the data under the Creative Commons. We will
update/change these license conditions as we see fit, and future updates
will apply back to this commit. The only reason we might change is if we
1) discover a license more appropriate for the data, or 2) decide that
our licensing is preventing the reasonable use of our data in research.

The programs (*.py) were updated to include a GPL header.

The entire text of the GPL license (v3.0) was included in the
repository.